Improved wrappers.py for auto code-generation
List of auto-generated files:
  fred_wrappers.h
  syscallsreal_helper.c
  wrapper_util.h
  wrapper_util.cpp
  fred_read_log.h

At this point, the work is complete but requires cleanup in wrappers.py
(since a bunch of code was written while commuting ;-)). The grammar is
explained at the top of wrappers.py file and should be complete as well.
There is room for improvement and so please let me know if you have any
comments.